+++ plugins/org.eclipse.actf.ai.tts.msp/src/org/eclipse/actf/ai/tts/msp/engine/SpEvent.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -19,9 +19,9 @@ import org.eclipse.actf.util.win32.MemoryUtil;
  */
 public class SpEvent {
 
-	private int address = MemoryUtil.GlobalAlloc(24);
+	private long address = MemoryUtil.GlobalAlloc(24);
 
-	public int getAddress() {
+	public long getAddress() {
 		return address;
 	}
